+++ app/routes/components/SortIndicator/index.tsx
@@ -41,7 +41,7 @@ export default function SortIndicatorPage() {
         </DocLayout.Content>
       )}
       {tabsValue === "source" && (
-        <DocLayout.Content>
+        <DocLayout.Content variant="files">
           <DocLayout.Files files={sourceCode} />
         </DocLayout.Content>
       )}