+++ src/components/SortIndicator/index.ts
@@ -1,8 +1,8 @@
-// SortIndicator from copy-ui @ 2025-11-15.
+// SortIndicator from copy-ui @ 2025-12-17.
 
 export type {
   SortIndicatorDirection,
   SortIndicatorProps,
 } from "./sort-indicator";
 export { SortIndicator } from "./sort-indicator";
-export { useDirection } from "./use-direction";
+export { useDirection, useDirections } from "./use-direction";