+++ src/components/SortIndicator/CHANGELOG.md
@@ -5,3 +5,5 @@
   the `SortIndicator` has its style.
 - 2025-11-15: Add the variant `compact`. The `size` prop is removed -- user
   should use CSS to set its size.
+- 2025-12-17: Add the new hook `useDirections`, which is useful to manage
+  multiple sort indicators.